+++ client/containers/SplashView.js
@@ -4,15 +4,15 @@ const connect           = require('react-redux').connect;
 // require components/files
 const SignIn            = require('../components/Splash/SignIn');
 const JumbotronInstance = require('../components/Splash/Jumbotron');
-const apiAction         = require('../actionCreators/apiActions');
+const userAction        = require('../actionCreators/userActions');
 const stateAction		= require('../actionCreators/stateActions');
 // require history to change routes
 const browserHistory    = require('react-router').browserHistory;
 
 const Splash = ({onSignIn, goSignup, logged }) => {
-	if(apiAction.getCookie("token")){
-		stateAction.SignInSuccess(apiAction.getCookie("token"))
-		apiAction.GetMyInfo();
+	if(userAction.getCookie("token")){
+		stateAction.SignInSuccess(userAction.getCookie("token"))
+		userAction.GetMyInfo();
 	}
 	return (
 		<div>
@@ -29,7 +29,7 @@ const Splash = ({onSignIn, goSignup, logged }) => {
 };
 
 const mapStateToProps = (state) => {
-	console.log(apiAction.getCookie("token"));
+	// console.log(userAction.getCookie("token"));
 	// ^^^ this console log returns each state in the reducers/reducer.js file
 	return {
 		logged: state.signin.logged
@@ -46,7 +46,7 @@ const mapDispatchToProps = (dispatch) => {
 			var username = e.target.username.value;
 			var password = e.target.password.value;
 			var body = {username: username, password: password};
-			dispatch(apiAction.SignIn(body)) // you have to call dispatch here even though you dispatch the action object in the api promise, idk why, check later
+			dispatch(userAction.SignIn(body)) // you have to call dispatch here even though you dispatch the action object in the api promise, idk why, check later
 		},
 		goSignup: () => {
 			browserHistory.push('/signup')
